Update README.md
I love what you have done, friend. Some of the words were misspelled and it could be a translation thing but take a look. The word developer had some letters cut off from being to long so I split it but then "I love to teach was short". What do you think? just close the Pull Request if you do not like it.

Oh and you already are a web developer so I added SR to the statement to show your still learning.